requirements.txt
# Core dependencies
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
dataclasses-json>=0.6.0

# LLM integration
openai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0

# RAG dependencies
chromadb>=0.4.0
tiktoken>=0.5.0
numpy>=1.24.0

# Git integration
gitpython>=3.1.0

# Message queue and protobuf
protobuf>=4.24.0
confluent-kafka>=2.3.0
aio-pika>=9.3.0
anyio>=4.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
ruff>=0.1.0